REF: Move old GROMACS module to new structure
