+++ openff/interchange/interop/gromacs/models/__init__.py
@@ -1,0 +1,1 @@
+"""Classes used to represent GROMACS state."""
