+++ openff/interchange/tests/interoperability_tests/internal/test_gromacs.py
@@ -8,6 +8,7 @@ from openff.toolkit.topology import Molecule, Topology
 from openff.toolkit.typing.engines.smirnoff import ForceField, VirtualSiteHandler
 from openff.toolkit.utils import get_data_file_path
 from openff.units import unit
+from openff.units.openmm import ensure_quantity
 from openff.utilities.testing import skip_if_missing
 from openmm import app
 from openmm import unit as openmm_unit
@@ -22,7 +23,7 @@ from openff.interchange.exceptions import (
     UnsupportedExportError,
     VirtualSiteTypeNotImplementedError,
 )
-from openff.interchange.interop.internal.gromacs import from_gro
+from openff.interchange.interop.gromacs._import._import import from_gro
 from openff.interchange.models import PotentialKey, TopologyKey
 from openff.interchange.tests import _BaseTest, needs_gmx
 
@@ -36,19 +37,20 @@ class TestGROMACSGROFile(_BaseTest):
             "tests/gromacs/unit_tests/angle10_vacuum/angle10_vacuum.gro",
         )
 
-        internal_coords = from_gro(file).positions.m_as(unit.nanometer)
-        internal_box = from_gro(file).box.m_as(unit.nanometer)
+        coords, box = from_gro(file)
 
         openmm_gro = openmm.app.GromacsGroFile(file)
-        openmm_coords = numpy.array(
-            openmm_gro.getPositions().value_in_unit(openmm_unit.nanometer),
+        openmm_coords = ensure_quantity(
+            openmm_gro.getPositions(),
+            "openff",
         )
-        openmm_box = numpy.array(
-            openmm_gro.getPeriodicBoxVectors().value_in_unit(openmm_unit.nanometer),
+        openmm_box = ensure_quantity(
+            openmm_gro.getPeriodicBoxVectors(),
+            "openff",
         )
 
-        assert numpy.allclose(internal_coords, openmm_coords)
-        assert numpy.allclose(internal_box, openmm_box)
+        assert numpy.allclose(coords, openmm_coords)
+        assert numpy.allclose(box, openmm_box)
 
     @skip_if_missing("intermol")
     def test_load_gro_nonstandard_precision(self):
@@ -56,7 +58,8 @@ class TestGROMACSGROFile(_BaseTest):
             "intermol",
             "tests/gromacs/unit_tests/lj3_bulk/lj3_bulk.gro",
         )
-        internal_coords = from_gro(file).positions.m_as(unit.nanometer)
+
+        coords = from_gro(file)[0].m_as(unit.nanometer)
 
         # OpenMM seems to assume a precision of 3. Use InterMol instead here.
         from intermol.gromacs.grofile_parser import GromacsGroParser
@@ -75,11 +78,11 @@ class TestGROMACSGROFile(_BaseTest):
             float,
         )
 
-        assert numpy.allclose(internal_coords, other_coords)
+        assert numpy.allclose(coords, other_coords)
 
         # This file happens to have 12 digits of preicion; what really matters is that
         # the convential precision of 3 was not used.
-        n_decimals = len(str(internal_coords[0, 0]).split(".")[1])
+        n_decimals = len(str(coords[0, 0]).split(".")[1])
         assert n_decimals == 12
 
     def test_vaccum_warning(self, sage):
